Move FileManager extensions
